pyproject.toml
[project]
name = "flow-state"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "neo4j>=5.28.1",
    "ipykernel>=6.29.5",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyyaml>=6.0.2",
    "discord-py>=2.5.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
